Rename files and fix READMEs
